+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_extends_keyword/examples/example.dart
@@ -1,0 +1,23 @@
+abstract class A {}
+
+class B extends A {} // LINT
+
+abstract class C {
+  int get a;
+}
+
+class D extends C { // LINT
+  // Implementation
+}
+
+abstract class E {
+  int get a;
+}
+
+abstract class G extends E { // LINT
+  String get b;
+}
+
+class K extends G { // LINT
+  // Implementation
+}